Methods for capturing stdout and stderr
