+++ lib/luffa.rb
@@ -1,1 +1,3 @@
 require 'luffa/patches/awesome_print'
+require 'luffa/patches/kernel'
+
