+++ spec/spec_helper.rb
@@ -3,27 +3,6 @@ require 'pry'
 require 'tmpdir'
 require 'open3'
 
-
-module Kernel
-  def capture_stdout
-    out = StringIO.new
-    $stdout = out
-    yield
-    return out
-  ensure
-    $stdout = STDOUT
-  end
-
-  def capture_stderr
-    out = StringIO.new
-    $stderr = out
-    yield
-    return out
-  ensure
-    $stderr = STDERR
-  end
-end
-
 # This file was generated by the `rspec --init` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
 # The generated `.rspec` file contains `--require spec_helper` which will cause this
